backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pytest
requests
httpx
